feat(query-builder): Move filter value suggestion and validation logic into own files
